test: consolidate shared Redis fixtures for integration tests
Before: each Redis-using suite (state/, exporters/, rate_limiting/) had its
own near-duplicate conftest.py + docker-compose.yml with its own copy of
is_redis_available / redis_service / DockerServices.

After: a single tests/integration/conftest.py owns the Redis lifecycle
(is_redis_available, redis_service, docker_ip, docker_compose_file) and a
single tests/integration/docker-compose.yml is the source of truth. Per-suite
conftests now contain only suite-specific fixtures (clear_state_keys,
clear_redis_streams, clear_rate_limit_keys, account_viewer, instruments).

Behavior preserved: redis_service reuses an already-running Redis on
localhost:6379 when available (matches the prior state/ behavior), and
starts one via docker-compose otherwise; GitHub Actions keeps using its
service container.

Also mark the new rate-limiting cross-loop tests with @pytest.mark.integration
so they run only under `just test-integration`, not `just test-ci`.
