--- tests/integration/exporters/docker-compose.yml
@@ -1,10 +1,0 @@
-services:
-  redis:
-    image: redis:latest
-    ports:
-      - "6379:6379"
-    healthcheck:
-      test: ["CMD", "redis-cli", "ping"]
-      interval: 1s
-      timeout: 3s
-      retries: 30
